feat!: 0.1.0-dev
this commit contains the very first working version of the from scratch rebuild. it contains the initialization of the project using uv, the system prompt, the tools that llm can use such as the rag, the shell and the web browser. it do not contains the datasets because i will use a nosql solution like mongodb instead of store everything in a directory like now.

BREAKING CHANGE: not working without datasets